requirements.txt
Django==1.7.4
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
celery==3.1.17
djangorestframework==3.0.4
kombu==3.0.24
pytz==2014.10
wsgiref==0.1.2
